Rename instrument modules to snake_case and update imports
