+++ AFL/automation/instrument/chessid3b.py
@@ -2,8 +2,8 @@ import gc
 import time
 import datetime
 from AFL.automation.APIServer.Driver import Driver
-from AFL.automation.instrument.ScatteringInstrument import ScatteringInstrument
-from AFL.automation.instrument.PySpecClient import PySpecClient
+from AFL.automation.instrument.scattering_instrument import ScatteringInstrument
+from AFL.automation.instrument.py_spec_client import PySpecClient
 import numpy as np # for return types in get data
 import h5py #for Nexus file writing
 import os
