+++ AFL/automation/instrument/file_camera.py
