+++ AFL/automation/instrument/network_camera.py
