+++ AFL/automation/instrument/nice_console.py
