+++ AFL/automation/instrument/nicos_script_client.py
