+++ AFL/automation/instrument/scattering_instrument.py
