+++ AFL/automation/instrument/scattering_interpolator.py
