+++ AFL/automation/instrument/sinqsans.py
@@ -2,7 +2,7 @@ import gc
 import time
 import datetime
 from AFL.automation.APIServer.Driver import Driver
-from AFL.automation.instrument.ScatteringInstrument import ScatteringInstrument
+from AFL.automation.instrument.scattering_instrument import ScatteringInstrument
 import numpy as np # for return types in get data
 import h5py #for Nexus file writing
 import os
