+++ AFL/automation/instrument/usb_camera.py
