+++ server_scripts/virtual_instrument/VirtualSpec_Data.py
@@ -7,7 +7,7 @@ except:
         print(f'Could not find NistoRoboto on system path, adding {os.path.abspath(Path(__file__).parent.parent)} to PYTHONPATH')
 import xarray as xr
 from AFL.automation.APIServer.APIServer import APIServer
-from AFL.automation.instrument.VirtualSpec_data import VirtualSpec_data
+from AFL.automation.instrument.virtual_spec_data import VirtualSpec_data
 from AFL.automation.APIServer.data.DataTiled import DataTiled
 
 ### local tiled server here
